chore(perf_hooks): extract timerify module
